팝업, 드롭다운 코드 분리
