+++ src/components/Editor/eventHandlers/attachDropdown.js
@@ -1,8 +1,39 @@
-import { Dropdown } from "../../Dropdown.js";
+import { createDebug } from "../../../shared/debug.js";
+import { Dropdown } from "../../Dropdown/dropdownElement.js";
+
+const debug = createDebug("Dropdown");
 
 export const enableDropdownFeature = ($editor, documentId) => {
     // 핸들러를 제대로 구현 할수록 더 많은 정보가 전달될 듯
-    const { $dropdown, displayDropdown } = Dropdown(documentId);
+    const $dropdown = Dropdown(documentId);
+
+    // mouseup 직후의 상태는 selection이 제거되어도 제거된 것을 인식하지 못 함.
+    // setTimeout이 왜 되는지 모르겠지만, 됨.
+    // keydown에서는 setTimeout 이어도 안 됨. keyup에서만 됨.
+
+    // TODO: 마우스로 드래그할 때는 좀 이상한 듯? 드래그했을 때 안 뜰 때가 있음. 확인 필요
+    // TODO: 역방향 드래그로 하면 인식이 안 됨.
+    // 아니, 왜 될 때가 있고 안 될 때가 있지?
+    // TODO: ESC 눌러야만 꺼지는데, 왜 ESC 누르면 꺼지는지 알아보기
+    const displayDropdown = () => {
+        setTimeout(() => {
+            // 1. Selection은 위치 확인 용도
+            // 무조건 textNode가 걸린다. slash를 입력했기 때문에.
+            const s = window.getSelection();
+            const oRange = s.getRangeAt(0); //get the text range
+            const oRect = oRange.getBoundingClientRect();
+            debug(s, oRange);
+
+            // 그냥 top, left만 주면 선택 영역과 popup이 겹침.
+            // height만큼 top에서 빼주면 딱 위에 붙음. 여기서 대충 한 10px 정도 더 빼주면 무난할 듯
+            // fadein-fadeout도 만들면 좋겠다.
+            // 그리고 그냥 css로 처리하는 게 좋을 듯? 물론 top/left는 js로 해야겠지만.
+            // TODO: 화면 꽤 하단에서 /를 누르면, 되게 올라간 위치에서 표시함.
+            $dropdown.style.removeProperty("display");
+            $dropdown.style.top = `${oRect.bottom + 4}px`;
+            $dropdown.style.left = `${oRect.left}px`;
+        }, 0);
+    };
 
     const openDropdownOnSlash = (e) => {
         // selection 반영을 위함.
